+++ src/Providers/RouteServiceProvider.php
@@ -50,6 +50,17 @@ class RouteServiceProvider extends ServiceProvider
         Route::bind('role', function ($value) {
             return Role::where('slug', $value)->firstOrFail();
         });
+
+        Route::bind('orchid_public', function ($path) {
+            $path = str_replace("..", "", $path);
+            $path = realpath(DASHBOARD_PATH . '/public/' . $path);
+
+            if($path !== false){
+                return $path;
+            }
+
+            abort(404);
+        });
     }
 
     /**